requirements.txt
absl_py~=0.6.1
pillow>=6.2.0
